+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changelog
 
+* [#32](https://github.com/dblock/slack-gamebot/issues/32): Players can express their frustration with `suck` - [@dblock](https://github.com/dblock).
 * [#26](https://github.com/dblock/slack-gamebot/issues/26): Players can `resign` a challenge - [@dblock](https://github.com/dblock).
 * [#65](https://github.com/dblock/slack-gamebot/issues/65): Captains can set per-team bot aliases with `set aliases <alias|none> ...` - [@dblock](https://github.com/dblock).
 * [#57](https://github.com/dblock/slack-gamebot/issues/57): Captains can turn team GIFs on/off with `set gifs on|off` - [@dblock](https://github.com/dblock).